requirements.txt
flask
requests
unidecode
bs4
Gunicorn
python-docx
tkinter
customtkinter